requirements.txt
streamlit>=1.36.0
openai>=1.40.0
Pillow>=10.0.0
extra-streamlit-components>=0.1.71

PyJWT>=2.8.0
cryptography>=42.0.0   # ← RS256/ES256 を使うなら
